#89 Renamed spline.h to spline_curve.h
