+++ graphics/spline_curve.h
